move testharness
