+++ packages/autotest/test/ClassPortalSpec.ts
@@ -4,7 +4,7 @@ import "mocha";
 import Config, {ConfigKey} from "../../common/Config";
 import Log from "@common/Log";
 
-import {Test} from "../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 import {AutoTestResult} from "../../common/types/AutoTestTypes";
 import {ContainerInput, ContainerOutput, ContainerState} from "../../common/types/ContainerTypes";
 import {AutoTestGradeTransport} from "../../common/types/PortalTypes";
