+++ packages/autotest/test/GitHubEventSpec.ts
@@ -4,7 +4,7 @@ import "mocha";
 
 import Config, {ConfigKey} from "../../common/Config";
 import Log from "../../common/Log";
-import {Test} from "../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 import {CommitTarget} from "../../common/types/ContainerTypes";
 import {DeliverablesController} from "../../portal/backend/src/controllers/DeliverablesController";
 import {PersonController} from "../../portal/backend/src/controllers/PersonController";
