+++ packages/portal/backend/test/CSVParserSpec.ts
@@ -2,7 +2,7 @@ import {expect} from "chai";
 import "mocha";
 
 import Log from "../../../common/Log";
-import {Test} from '../../../common/TestHarness';
+import {Test} from '@common/test/TestHarness';
 import {GradesController} from "../src/controllers/GradesController";
 import {CSVParser} from "../src/server/common/CSVParser";
