+++ packages/portal/backend/test/ClasslistAgentSpec.ts
@@ -2,7 +2,7 @@ import {expect} from "chai";
 import "mocha";
 
 import Log from "../../../common/Log";
-import {Test} from '../../../common/TestHarness';
+import {Test} from '@common/test/TestHarness';
 import {ClasslistChangesTransportPayload} from "../../../common/types/PortalTypes";
 import {ClasslistAgent} from "../src/server/common/ClasslistAgent";
