+++ packages/portal/backend/test/controllers/DatabaseControllerSpec.ts
@@ -2,7 +2,7 @@ import {expect} from "chai";
 import "mocha";
 import Config, {ConfigKey} from "../../../../common/Config";
 import Log from "../../../../common/Log";
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 
 import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {PersonKind} from "../../src/Types";
