+++ packages/portal/backend/test/controllers/DeliverablesControllerSpec.ts
@@ -6,7 +6,7 @@ import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {DeliverablesController} from "../../src/controllers/DeliverablesController";
 import {Deliverable} from "../../src/Types";
 
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 
 import "../GlobalSpec";
 import "./DatabaseControllerSpec"; // run first
