+++ packages/portal/backend/test/controllers/GitHubActionSpec.ts
@@ -3,7 +3,7 @@ import "mocha";
 
 import Config, {ConfigKey} from "@common/Config";
 import Log from "@common/Log";
-import {Test} from "@common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 import Util from "@common/Util";
 
 import {DatabaseController} from "@backend/controllers/DatabaseController";
