+++ packages/portal/backend/test/controllers/GradeControllerSpec.ts
@@ -1,7 +1,7 @@
 import {expect} from "chai";
 import "mocha";
 
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 import {AutoTestGradeTransport} from "../../../../common/types/PortalTypes";
 import {GradePayload} from "../../../../common/types/SDMMTypes";
