+++ packages/portal/backend/test/controllers/ResultControllerSpec.ts
@@ -1,6 +1,6 @@
 import {expect} from "chai";
 import "mocha";
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 import {GradeReport} from "../../../../common/types/ContainerTypes";
 
 import {AutoTestResultTransport} from "../../../../common/types/PortalTypes";
