+++ packages/portal/backend/test/controllers/TeamControllerSpec.ts
@@ -2,7 +2,7 @@ import {expect} from "chai";
 import "mocha";
 
 import Log from "../../../../common/Log";
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 
 import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {PersonController} from "../../src/controllers/PersonController";
