+++ packages/portal/backend/test/controllers/TestGitHubActions.ts
@@ -1,7 +1,7 @@
 // do not use module imports (@common, @backend) in this file
 import Log from "../../../../common/Log";
 import Config, {ConfigKey} from "../../../../common/Config";
-import {Test} from "../../../../common/TestHarness";
+import {Test} from "@common/test/TestHarness";
 
 import {BranchRule, GitPersonTuple, GitRepoTuple, GitTeamTuple, Issue} from "../../src/controllers/GitHubController";
 import {TeamController} from "../../src/controllers/TeamController";
