+++ testOutput/index.html
@@ -3,14 +3,13 @@
 	<title>Classy CI Artifacts</title>
 </head>
 <body style="font-size: 2em;">
-	
 	<!-- These resources are created by the CI environment and are not actually checked into version control -->
-	<center>
+	<div style="text-align: center;">
 		<h2>Classy CI Artifacts</h2>
-		<a href="backend/test/index.html">Backend Test Results</a>
+		<br/><a href="backend/test/index.html">Backend Test Results</a>
 		<br/><a href="backend/coverage/index.html">Backend Coverage Report</a>
 		<br/><a href="autotest/test/index.html">Autotest Test Results</a>
 		<br/><a href="autotest/coverage/index.html">Autotest Coverage Report</a>
-	</center>
+	</div>
 </body>
 </html>
